+++ packages/jest-cli/src/init/index.ts
@@ -13,8 +13,8 @@ import {constants} from 'jest-config';
 import {tryRealpath} from 'jest-util';
 import defaultQuestions, {testScriptQuestion} from './questions';
 import {MalformedPackageJsonError, NotFoundPackageJsonError} from './errors';
-import generateConfigFile from './generate_config_file';
-import modifyPackageJson from './modify_package_json';
+import generateConfigFile from './generateConfigFile';
+import modifyPackageJson from './modifyPackageJson';
 import type {ProjectPackageJson} from './types';
 
 const {