+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixes
 
 - `[expect]` Stop modifying the sample in `expect.objectContaining()` ([#10711](https://github.com/facebook/jest/pull/10711))
+- `[jest-resolve-dependencies]` Resolve mocks as dependencies ([#10713](https://github.com/facebook/jest/pull/10713))
 
 ### Chore & Maintenance